Working dockerfile.
We make use of a two-stage docker build process.

The first build stage installs the OCaml libraries
to compile the server. It also stores the dynamic
dependencies of the server thanks to the opam depext
command.

The second and final stage copies the server executable
and installs its dynamic dependencies. It is saved as the
resulting image, running the server.