Change expiration date properly
I mistakenly wrote 55*10 instead of 60*8 since it's not 55 minutes instead of 60 but 8 minutes instead of 10

A safe way would be, if you want me to try it, to 

> date && curl -I https://api.github.com | grep -Fi "date"

Read the two dates, compute the difference in time and create the expected expiration date for the token accordingly. This way, it doesn't matter if your computer is way out of sync, you'll always be able to provide GitHub with the proper dates.