Improve minimization logic around scripts and files
We now assume that `@coqbot minimize` is followed by a Coq file, not a
bash script, unless the user begins with ` ```sh` or ` ```bash` or the
script begins with `#!`.

Additionally, we also allow the user to say `@coqbot minimize
[text](url)`, and use the `handle-web-file.sh` script in the
run-coq-bug-minimizer to automatically download `url` and automagically
figure out how to build run coq on it.  (Currently there's support for
uncompressing archives, running single .v files, renaming text files to
.v if there's no archive, finding _CoqProject, and using autogen,
configure, and make to build archives if there's a single makefile.)

Fixes #256