Add support for mirroring GitHub repositories to GitLab.
For now, the list of repositories is hardcoded. If we want to make
this configurable per repository, we need to ensure that the GitLab
repository does belong to the claimed GitHub owner, e.g., by checking
the configuration file both on GitHub and GitLab, or by requiring the
use of a specific GitLab project description with a backlink to the
GitHub repository.

Co-authored-by: Erik Martin-Dorel <erik.martin-dorel@irit.fr>